Clean up
- Generic opus_set_complexity() wrapper around macro of same name.
- Cast C.malloc argument with C.size_t for cross-architecture support.

+++ media_sources_linux.go
@@ -189,7 +189,7 @@ func (as *ALSAAudioSource) capture() {
 			return
 		default:
 			// Capture
-			out := C.malloc(C.uint(bytesPerSample * numChannels * numFrames))
+			out := C.malloc(C.size_t(bytesPerSample * numChannels * numFrames))
 			n := C.snd_pcm_readi(
 				as.handle,
 				unsafe.Pointer(out),
